fix(userapp): update 管理面组——ports live 回退+发布并发 409+pingora 失败恢复+查询边角
ports 回退（P1）：update 不带 ports 此前静默清空全部端口（SSA 清 container
ports + cleanup_orphan_port_resources 删 HTTPRoute/NodePort → 对外入口全断），
注释声称"与 ports 回退一致"实际从未实现。ContainerSpecSnapshot 增 ports
字段：K8s 从 container.ports + port-expose 注解还原 name/port/expose_type
（strip_prefix 未持久化读回 None，取舍见字段文档）；Docker 从 ExposedPorts
尽力还原（expose_type 恒 Http 兜底，保住端口号不丢是主要目标）。

update 与发布并发（P1）：不持发布锁可与 delete 并发造成"删除成功但
Deployment 复活"。按业务决策改为 try_lock——发布进行中立即 409 让调用方
稍后重试，不排队傻等 activate 最长 30 分钟的就绪窗口；成功路径尾部对称
drop+清理锁条目（对齐 delete_app）。

pingora 失败恢复（P1）：update 先注销网关路由再 patch 集群，patch 失败
路由不恢复——应用还在运行但 /proxy 入口 404。对齐 delete_app 的失败恢复
分支（current.ports 的 HTTP 端口 + 旧 pod_ip 重注册）。

Docker 快照正确性：command 不再读回（inspect 的 Config.cmd 无法区分用户
显式设置与镜像 CMD 固化，读回会把旧镜像 CMD 钉死到换镜像后的新容器——
缺省走新镜像自己的 CMD 才是升级语义）；docker_memory_to_quantity 无损化
（此前就近向下取整，1.5Gi 部分更新回退后永久缩水成 1Gi）。

查询边角（P2）：分页对齐 query_storage/publish tasks 口径（page<1 或
page_size∉[1,100] → 400；u64 中间量防 debug 构建溢出 panic/环绕错页）；
补 created_at 排序（metadata join，缺元数据排最后）；非法 sort_by → 400
（此前静默不排但 reverse 仍执行，半生效）；name 过滤改模糊匹配（对齐
"按名称模糊搜索"的契约注释）；publish tasks 内存分页乘法同改 u64。

测试：ports 回退（含 TCP/Http/strip）+ 发布锁 409 + 非法分页/排序 400；
QueryAppsRequest 补 Default。

+++ crates/rcoder/src/userapp_publish/store.rs
@@ -249,7 +249,9 @@ impl PublishTaskStore {
         let total = snapshots.len() as u64;
         let page = page.max(1);
         let page_size = page_size.clamp(1, 100);
-        let start = ((page - 1) * page_size) as usize;
+        // u64 中间量防溢出（handler 已校验 page>=1，但无上限——debug 构建下 u32 乘法
+        // 溢出 panic；极端页码截断为越界空页）
+        let start = ((page as u64 - 1) * page_size as u64) as usize;
         let end = (start + page_size as usize).min(snapshots.len());
         let items = if start < snapshots.len() {
             snapshots[start..end].to_vec()
